jobsub tools v1_2v

+++ jobsub_tools/prd/jobsub_tools/v1_2/Linux-2/bin/unsetup_condor.csh
@@ -3,4 +3,7 @@
 setenv source_me `${JOBSUB_TOOLS_DIR}/bin/unsetup_condor_sh.py`
 
 source $source_me
-/bin/rm $source_me
+if ( "$?REMOVE_JOBSUB_UPS_SOURCE" = "0" ) then
+    /bin/rm $source_me
+endif
+
