+++ jobsub_tools/prd/jobsub_tools/v1_2/Linux-2/bin/unsetup_condor.sh
@@ -3,5 +3,9 @@
 export source_me=`${JOBSUB_TOOLS_DIR}/bin/unsetup_condor_sh.py`
 
 source $source_me
-/bin/rm $source_me
+if [ "$REMOVE_JOBSUB_UPS_SOURCE" = "" ] 
+then
+        /bin/rm $source_me
+fi
+
 unset source_me
